Struct uhlc::HLC

pub struct HLC { /* fields omitted */ }

An Hybric Logical Clock generating [Timestamp]s

Implementations

impl HLC

pub fn with_clock(id: Vec<u8>, clock: fn() -> NTP64) -> HLC

Create a new [HLC] with an id (must be unique) and a clock (a function returning the current physical time as an [NTP64]).

Examples

use uhlc::HLC;
 
let hlc = uhlc::HLC::with_clock(vec![0x01, 0x02, 0x03], uhlc::system_time_clock);
println!("{}", hlc.new_timestamp().await);

pub fn with_system_time(id: Vec<u8>) -> HLC

Create a new [HLC] with an id (must be unique) and using [system_time_clock()] as physical clock.

Examples

use uhlc::HLC;
 
let hlc = uhlc::HLC::with_system_time(vec![0x01, 0x02, 0x03]);
println!("{}", hlc.new_timestamp().await);

pub async fn new_timestamp<'_>(&'_ self) -> Timestamp

Generate a new [Timestamp].

This timestamp is unique in the system and is always greater than the latest timestamp generated by the HLC and than the latest incoming timestamp that was used to update this [HLC] (using [HLC::update_with_timestamp()]).

Examples

use uhlc::HLC;
 
let hlc = uhlc::HLC::default();
let ts1 =  hlc.new_timestamp().await;
let ts2 =  hlc.new_timestamp().await;
assert!(ts2 > ts1);

pub async fn update_with_timestamp<'_, '_>(
&'_ self,
timestamp: &'_ Timestamp
) -> Result<(), String>

Update this [HLC] with a [Timestamp].

Typically, this timestamp should have been generated by another HLC. If the timestamp exceeds the current time of this HLC by more than [DELTA_MS] an [Err] is returned.

Examples

use uhlc::HLC;
 
let hlc1 = HLC::default();
 
// update the HLC with a timestamp incoming from another HLC
// (typically remote, but not in this example...)
let hlc2 = HLC::default();
let other_ts = hlc2.new_timestamp().await;
if ! hlc1.update_with_timestamp(&other_ts).await.is_ok() {
    println!(r#"The incoming timestamp would make this HLC
             to drift too much. You should refuse it!"#);
}
 
let ts = hlc1.new_timestamp().await;
assert!(ts > other_ts);